moved folder up from txmgr to common
